# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Power Vital API",
    "version": "1.0.0",
    "description": "Power Vital e-commerce + MLM platform backend.\n\n## Languages\nAll user-facing strings live in three locales: **TR** (primary), **RU**, **KG**.\nThe active locale is negotiated via the `Accept-Language` header and a `lang`\nquery param. AI-generated translations fill missing locale fields automatically.\n\n## Auth\nMutating endpoints require a `Bearer` JWT in the `Authorization` header.\nAdmin-only routes additionally check the `role` claim.\n\n## Rate limiting\nAll `/api/*` routes are rate-limited to 150 req/min per IP. Burst over\nthe limit returns 429 with `Retry-After`.\n\n## Errors\nAll errors return `{ error: string, issues?: ValidationIssue[] }` where\n`ValidationIssue` has the shape `{ path, message, code }`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://powervital.kg",
      "description": "Production (Kyrgyzstan)"
    },
    {
      "url": "https://powervital.org",
      "description": "Production (intl alias)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/finance/exchange-rate": {
      "get": {
        "tags": [
          "Exchange Rate"
        ],
        "description": "Get the current USD→KGS rate (?refresh=true forces a fetch)",
        "responses": {
          "200": {
            "description": "Rate"
          },
          "404": {
            "description": "No rate available"
          }
        }
      }
    },
    "/api/v1/finance/exchange-rate/refresh": {
      "post": {
        "tags": [
          "Exchange Rate"
        ],
        "description": "Manually trigger a rate fetch (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "502": {
            "description": "Provider failed"
          }
        }
      }
    },
...
    "/api/v1/errors/report": {
      "post": {
        "tags": [
          "Errors"
        ],
        "description": "Report a client-side error. Open to anonymous, rate-limited.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "stack": {
                    "type": "string",
                    "maxLength": 32768
                  },
                  "source": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "phase": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "route": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "locale": {
                    "type": "string",
                    "maxLength": 8
                  },
                  "context": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "clientTimestamp": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "message",
                  "source"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Error persisted"
          },
          "400": {
            "description": "Invalid payload"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/v1/errors/recent": {
      "get": {
        "tags": [
          "Errors"
        ],
        "description": "Admin: recent client errors. Pass ?resolved=true to include resolved.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "resolved",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent errors"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Admin only"
          }
        }
      }
    },
    "/api/v1/errors/{id}/resolve": {
      "post": {
        "tags": [
          "Errors"
        ],
        "description": "Admin: mark a client error as resolved.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolvedNote": {
                    "type": "string",
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
...
}